requirements.txt
django-model-utils
Django==2.0.6
pbr==4.0.4
pytz==2018.4
six==1.11.0
stevedore==1.28.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
